// app/api/providers/elasticsearch/healthRecordSummaryIndex/interfaces/ESHealthRecordSummaryInterface.tsx
export interface ESHealthRecordSummaryInterface {
  health_id: string;
  active: boolean | null;
  bin_brn: string | null;
  created_at: any;
  created_client_id: Number | null;
  created_facility_id: Number | null;
  national_id: string | null;
  updated_at: any
  updated_facility_id: Number | null;
}